114- feat(QrCodeScanner): Move QrCodeScanner to utils in the App, decrease padding background of qr code image
