+++ packages/ui_library/lib/widgets/u_qr_code/u_qr_code_export.dart
@@ -1,2 +1,1 @@
 export 'u_qr_code.dart';
-export 'u_qr_code_scanner.dart';
